% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{available_languages}
\alias{available_languages}
\title{Print possible search languages}
\usage{
available_languages()
}
\description{
Prints a list of languages that write_search() can write searches in.
}
\examples{
available_languages()
}
