% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_dfm}
\alias{BBWO_dfm}
\title{Example document-feature matrix}
\format{A document-feature matrix with 36 documents and 79 features}
\usage{
BBWO_dfm
}
\description{
A document-feature matrix containing a subset of example
Black-backed Woodpecker (Picoides arcticus) articles
and keywords.
}
\keyword{datasets}
