% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_graph}
\alias{BBWO_graph}
\title{Example keyword co-occurrence network}
\format{An igraph object}
\usage{
BBWO_graph
}
\description{
An example of a keyword co-occurrence network for a subset of
articles about Black-backed Woodpeckers (Picoides arcticus).
}
\keyword{datasets}
