% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_grouped_keywords}
\alias{BBWO_grouped_keywords}
\title{Grouped keywords for the BBWO vignette}
\format{A data frame with 429 rows of 2 variables:
\describe{
  \item{group}{concept group to which a term belongs}
  \item{term}{the term or phrase output by litsearchr}
}}
\usage{
BBWO_grouped_keywords
}
\description{
A dataset containing the keyword groupings used in the
vignette about Black-backed Woodpecker occupancy of
post-fire forest systems.
}
\keyword{datasets}
