% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_import}
\alias{BBWO_import}
\title{Example imported articles}
\format{A data frame with 36 rows of 18 variables for each article:
\describe{
 \item{id}{A unique ID}
 \item{text}{The title, abstract, and/or keywords combined}
 \item{title}{The title}
 \item{abstract}{The abstract}
 \item{keywords}{The author- or database-tagged keywords}
 \item{methods}{Database-tagged methods}
 \item{type}{Type of article (e.g. thesis, journal, book chapter)}
 \item{authors}{Authors of the article}
 \item{affiliation}{Author institutional affiliation}
 \item{source}{Journal or source of publication}
 \item{year}{Year of publication}
 \item{volume}{Volume of publication, if applicable}
 \item{issue}{Issue of publication, if applicable}
 \item{startpage}{The first page of the article}
 \item{endpage}{The last page of the article}
 \item{doi}{The DOI of the article}
 \item{language}{Original language of publication}
 \item{database}{The database from which the article was exported}
 }}
\usage{
BBWO_import
}
\description{
A data frame containing a subset of example imported
Black-backed Woodpecker (Picoides arcticus) articles
prior to deduplication.
}
\keyword{datasets}
