% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_similar_grouped}
\alias{BBWO_similar_grouped}
\title{Similar terms from example}
\format{An object of class \code{data.frame} with 922 rows and 2 columns.}
\usage{
BBWO_similar_grouped
}
\description{
A dataset of terms marked as similar when we were
building an example search strategy.
}
\keyword{datasets}
