% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{check_recall}
\alias{check_recall}
\title{Check the recall of a search strategy}
\usage{
check_recall(true_hits, retrieved)
}
\arguments{
\item{true_hits}{a character vector of titles for articles that should be returned}

\item{retrieved}{a character vector of titles for articles returned by a search}
}
\value{
a table of the best match for each true title from the search results along with a title similarity score
}
\description{
Checks a list of known articles against the results of a search to determine which gold standard hits are retrieved.
}
\examples{
check_recall(
  true_hits = c("Picoides arcticus"),
  retrieved = c("Picoides tridactylus",
                         "Seiurus aurocapilla")
)
}
