% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_testing.R
\name{check_search_strategy}
\alias{check_search_strategy}
\title{Check search strategy precision and sensitivity}
\usage{
check_search_strategy(relevant_studies, retrieved_studies, min_sim = 0.8,
  print_sim_min = 0.5)
}
\arguments{
\item{relevant_studies}{a character vector of article titles deemed relevant to a review}

\item{retrieved_studies}{a character vector of article titles retrieved by a search (with duplicates removed)}

\item{min_sim}{the minimum similarity for an article to be counted as being retrieved}

\item{print_sim_min}{the minimum similarity for an article to be returned for manual review}
}
\value{
a list with two objects: a table of search metrics (precision, number needed to process, and sensitivity) and a data frame of all titles whose similarity score is greater than print_sim_min
}
\description{
NOTE: this function takes a long time for large searches; use check_recall() along with search_performance() for large datasets. Calculates precision and sensitivity of a search given a set of known articles that should be included in the search and the complete and deduplicated results of conducting a search.
}
