% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Remove duplicate studies and punctuation}
\usage{
clean_dataset(df, chars = 40)
}
\arguments{
\item{df}{The dataset you want to deduplicate.}

\item{chars}{How many characters you want to include for deduplication (default is 40).}
}
\value{
The cleaned dataset.
}
\description{
This function removes entries from the database that are duplicated articles (e.g. articles that were returned by two different databases) based on the first 40 characters of the title. It also replaces all miscellaneous punctuation marks used to separate keywords and replaces them with a semicolon so that keywords properly separate in later steps.
}
