% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{clean_keywords}
\alias{clean_keywords}
\title{Remove duplicate studies and punctuation}
\usage{
clean_keywords(df)
}
\arguments{
\item{df}{a data frame from import_scope() to deduplicate}
}
\value{
a data frame with keyword punctuation standardized
}
\description{
Replaces all miscellaneous punctuation marks used to separate keywords and replaces them with a semicolon so that keywords properly separate in later steps.
}
