% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_alphas}
\alias{color_alphas}
\title{Color transparency conversion numbers}
\format{A data frame with 256 rows of 2 variables:
\describe{
  \item{code}{the alpha code}
  \item{intensity}{the percent color transparency to which a code corresponds}
}}
\usage{
color_alphas
}
\description{
A dataset containing conversion codes to go from a transparency
setting to an alpha code appended to a hex code.
}
\keyword{datasets}
