% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{condense_network}
\alias{condense_network}
\title{Condense a keyword co-occurrence network by removing irrelevant terms}
\usage{
condense_network(full_graph, rejected_terms,
  previous_rejected_terms = NULL)
}
\arguments{
\item{full_graph}{the full graph produced with create_network()}

\item{rejected_terms}{a character vector of terms rejected in the manual review stage}

\item{previous_rejected_terms}{a list of character vectors rejected in previous manual reviews. If this is the first iteration, it should be set to NULL.}
}
\value{
the full graph with the rejected term nodes deleted
}
\description{
Eliminates the rejected terms from the manual keyword review and rebuilds the network with the eliminated terms removing, making it recentralize on the terms marked as relevant in the manual stage.
}
\examples{
condense_network(full_graph=litsearchr::BBWO_graph, rejected_terms=c("actual increase"))
}
