% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{create_network}
\alias{create_network}
\title{Create a keyword co-occurrence network}
\usage{
create_network(search_dfm, min_studies = 3, min_occurrences = 3)
}
\arguments{
\item{search_dfm}{a document-feature matrix created with create_dfm()}

\item{min_studies}{the minimum number of studies a term must occur in to be included}

\item{min_occurrences}{the minimum total number of times a term must occur (counting repeats in the same document)}
}
\value{
an igraph weighted graph
}
\description{
Creates a keyword co-occurrence network from an adjacency matrix trimmed to remove rare terms.
}
