% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{create_subjects}
\alias{create_subjects}
\title{Create searchable study subjects}
\usage{
create_subjects(df, keys = TRUE, titles = TRUE, abstracts = TRUE,
  models = FALSE)
}
\value{
The same dataset with an additional column vector of the concatenated subject.
}
\description{
Combines titles, abstracts, keywords, and/or other aspects of the study (e.g. modeling approach) from metadata to search for untagged keywords.
}
