% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_stopwords}
\alias{custom_stopwords}
\title{Stopwords to remove from keywords}
\format{A character vector with 593 entries}
\usage{
custom_stopwords
}
\description{
A dataset of common terms to use as stopwords when tokenizing texts.
}
\keyword{datasets}
