% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_list}
\alias{database_list}
\title{Databases that can be imported with litsearchr}
\format{A database with 24 rows of 3 variables
\describe{
  \item{Platform}{the platform used to access a database}
  \item{Database}{the database accessed}
  \item{Download_method}{the way results must be downloaded for litsearchr to recognize the format}
}}
\usage{
database_list
}
\description{
A dataset containing the platforms, databases, and download methods
that litsearchr recognizes and can import.
}
\keyword{datasets}
