% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{deduplicate}
\alias{deduplicate}
\title{Remove duplicate articles}
\usage{
deduplicate(df, doc_sim = 0.85, title_sim = 0.95, mean_sim = 0.8,
  title_method = "tokens")
}
\arguments{
\item{df}{a data frame created with import_scope()}

\item{doc_sim}{the minimum similarity between two abstracts to be marked as duplicated}

\item{title_sim}{the minimum similarity between two titles to be marked as duplicated}

\item{mean_sim}{the minimum mean similarity of abstract and title similarity to be marked as duplicated}
}
\value{
a data frame with duplicates removed
}
\description{
Uses similarity of tokenized abstracts and titles to detect duplicates and remove them from the dataset.
}
