% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{detect_database}
\alias{detect_database}
\title{Detects which database a search is from}
\usage{
detect_database(df)
}
\arguments{
\item{df}{an exported dataset from any supported database}
}
\value{
a character vector with the name of the database or an error that the database was not identified
}
\description{
Uses the column names from databases to identify which database a search is from. This function can detect searches done in Web of Science databases (BIOSIS, Zoological Record, or MEDLINE), Scopus, and any EBSCO-indexed database.
}
