% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract potential keywords from abstracts and titles}
\usage{
extract_terms(
  text = NULL,
  keywords = NULL,
  method = c("fakerake", "tagged"),
  min_freq = 2,
  ngrams = TRUE,
  min_n = 2,
  max_n = 5,
  stopwords = NULL,
  language = "English"
)
}
\arguments{
\item{text}{A character object of text from which to extract terms}

\item{keywords}{A character vector of keywords tagged by authors and/or databases if using method="tagged"}

\item{method}{The method of extracting keywords; options are fakerake (a quick implementation similar to Rapid Automatic Keyword Extraction), or tagged for author-tagged keywords}

\item{min_freq}{Numeric: the minimum occurrences of a potential term}

\item{ngrams}{Logical: should litsearchr only extracts phrases with word count greater than a specified n?}

\item{min_n}{Numeric: the minimum length ngram to consider}

\item{max_n}{Numeric: the maximum length ngram to consider}

\item{stopwords}{A character vector of stopwords.}

\item{language}{A string indicating the language of input data to use for stopwords if none are supplied.}
}
\value{
Returns a character vector of potential keyword terms.
}
\description{
Extracts potential keyword terms from text (e.g. titles and abstracts)
}
\examples{
extract_terms(
  text = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  method = "fakerake",
  min_freq=1,
  min_n=2
)
}
