% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{find_cutoff}
\alias{find_cutoff}
\title{Find node cutoff strength}
\usage{
find_cutoff(
  graph,
  method = c("changepoint", "cumulative"),
  percent = 0.8,
  knot_num = 3,
  imp_method = "strength"
)
}
\arguments{
\item{graph}{An igraph graph object}

\item{method}{the cutoff method to use, either "changepoint" or "cumulative"}

\item{percent}{if using method cumulative, the total percent of node strength to capture}

\item{knot_num}{if using method changepoint, the number of knots to identify}

\item{imp_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a vector of suggested node cutoff strengths
}
\description{
Find the minimum node strength to use as a cutoff point for important nodes.
}
\details{
The changepoint fit finds tipping points in the ranked order of node strengths to use as cutoffs. The cumulative fit option finds the node strength cutoff point at which a certain percent of the total strength of the graph is captured (e.g. the fewest nodes that contain 80\% of the total strength).
}
\examples{
dfm <- create_dfm(
  elements = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  features = c("occupancy", "variation", "black-backed woodpecker", "burn")
)

my_network <- create_network(search_dfm = as.matrix(dfm),
                             min_studies = 1,
                             min_occ = 1)

find_cutoff(my_network,
            method = "cumulative",
            percent = .9)

find_cutoff(my_network,
            method = "changepoint",
            knot_num = 3)
}
