% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{get_descriptors}
\alias{get_descriptors}
\title{Search study subjects for keywords}
\usage{
get_descriptors(df, keywords, stemming = FALSE, minlength = 5,
  exact = FALSE)
}
\arguments{
\item{df}{The dataset.}

\item{keywords}{A vector of unique keywords in the sample.}

\item{stemming}{Defaults to false, but can search for stemmed versions of keywords.}

\item{minlength}{The minimum length for a word stem to be allowed to stem. Do not set to less than 3. For example, stemming "relative" results in "rel" which will also pick up unrelated words like relationship, religion, relict, etc...}

\item{exact}{Only searches for complete phrases and not partial words. If set to \code{TRUE}, the keyword "fire" will only pick up that exact phrase, but will not pick up "wildfire".}
}
\value{
The dataset with an additional column vector of "descriptors" (keywords and newly generated keywords from titles and abstracts).
}
\description{
Searches the study subjects, specified in \code{create_subjects}, to add additional keywords that were not tagged by the authors but appear in the title or abstract.
}
