% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{get_keywords}
\alias{get_keywords}
\title{Extract potential keywords}
\usage{
get_keywords(reduced_graph, savekeywords = TRUE, makewordle = TRUE)
}
\arguments{
\item{reduced_graph}{a reduced graph with only important nodes created with reduce_grah()}

\item{savekeywords}{if TRUE, saves the keywords to a plain text file}

\item{makewordle}{if TRUE, creates a wordcloud image of the important keywords sized relative to node strength}
}
\value{
a list of potential keywords to consider
}
\description{
Extracts keywords identified as important.
}
