% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{get_similar_terms}
\alias{get_similar_terms}
\title{Retrieves terms similar to included terms}
\usage{
get_similar_terms(grouped_terms, graph, considered_terms = NULL,
  ignore_terms = NULL)
}
\arguments{
\item{grouped_terms}{a list of character vectors with terms}

\item{graph}{a graph object of the full keyword co-occurrence}

\item{considered_terms}{a character vector of terms you already considered and rejected}

\item{ignore_terms}{a character vector of unigrams that should NOT be used to retrieve similar terms}
}
\value{
a named number vector of node strengths
}
\description{
Given a list of terms selected for inclusion, returns other terms from the network that are similar.
}
