% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Retrieve stopwords for a given language}
\usage{
get_stopwords(language = "English")
}
\arguments{
\item{language}{A character vector containing the name of the language for which to retrieve stopwords. Defaults to "English"}
}
\value{
Returns a character vector of stopwords.
}
\description{
This function retrieves stopwords to use for a specified language.
}
\examples{
get_stopwords("English")
}
