% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{get_tokens}
\alias{get_tokens}
\title{Remove stopwords from text}
\usage{
get_tokens(text, language = "English")
}
\arguments{
\item{text}{A character vector containing text from which to remove stopwords.}

\item{language}{A string indicating the language of the text.}
}
\value{
Returns the input text with stopwords removed.
}
\description{
Removes stopwords from text in whichever language is specified.
}
\examples{
get_tokens("On the Origin of Species", language="English")
}
