% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{get_unique}
\alias{get_unique}
\title{Find unique keywords}
\usage{
get_unique(df, singular = TRUE, minlength = 4, maxlength = 100)
}
\arguments{
\item{df}{The dataset of titles, abstracts, and keywords.}

\item{singular}{If \code{TRUE}, plural keywords are made singular (e.g. animals and animal count as the same word).}
}
\value{
A vector of unique keyword from all studies in dataset.
}
\description{
Many keywords are repeated across studies; only searching for unique keywords reduces computation time in later steps.
}
