% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{import_naive}
\alias{import_naive}
\title{Import results of naive searches}
\usage{
import_naive(file, database = c("Scopus", "ZooRec", "BIOSIS"))
}
\arguments{
\item{file}{This function assumes you have exported full records to a .csv.}

\item{database}{Which of the three supported databases your records are from.}
}
\value{
A dataset with just the relevant columns, in the same order, so that results from the three databases can be combined with rbind.
}
\description{
This function takes results of searches in BIOSIS, Zoological Record, or Scopus and converts them to the format used in this package. If you have keywords from a different database, make sure there are columns for id, title, abst (Abstract), key (keywords), and mods (other options, in this case, models).
}
