% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{import_results}
\alias{import_results}
\title{Import results of a search}
\usage{
import_results(directory = NULL, file = NULL, verbose = TRUE)
}
\arguments{
\item{directory}{a path to a directory containing search results to import}

\item{file}{a file of search results to import}

\item{verbose}{if TRUE, prints status updates}
}
\value{
a data frame of assembled search results
}
\description{
Given a file or directory containing bibliographic files, import and assemble search results
}
\details{
If a .bib or .ris file returns an error, there is likely a character encoding or formatting issue with the file. Please try uploading your .bib or .ris file to a reference manager (e.g. Zotero) and re-exporting it as a .bib or .ris, which often clears up formatting issues that import_results cannot detect.
}
