% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{import_scope}
\alias{import_scope}
\title{Import results of a scoping search}
\usage{
import_scope(directory, remove_duplicates = TRUE, clean_dataset = TRUE,
  save_full_dataset = FALSE)
}
\arguments{
\item{directory}{the full path to the directory in which the searches are saved}

\item{remove_duplicates}{if TRUE, removes duplicates based on document similarity}

\item{clean_dataset}{if TRUE, removes excess punctuation and standardizes keywords}

\item{save_full_dataset}{if TRUE, saves a .csv of the full dataset in the working directory}
}
\value{
a data frame of all the search results combined
}
\description{
Imports the results of a scoping search, combines them into a single dataset, and (optionally) removes duplicate hits based on document similarity. Duplicates can be removed subsequently with custom similarity cutoffs using deduplicate() on the full dataset.
}
