% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importable_databases}
\alias{importable_databases}
\title{Databases that can be imported with litsearchr}
\format{A list with 6 entries:
\describe{
  \item{BIOSIS CITATION INDEX}{}
  \item{EBSCO databases}{}
  \item{Scopus}{}
  \item{Zoological Record}{}
  \item{Web of Science - All Databases}{}
  \item{MEDLINE}{}
}}
\usage{
importable_databases
}
\description{
A dataset containing the database signatures (i.e. the column names
that are automatically exported with the full report option) for
databases that litsearchr can import.
}
\keyword{datasets}
