% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{language_code}
\alias{language_code}
\title{Get short language codes}
\usage{
language_code(language)
}
\arguments{
\item{language}{A character vector containing the name of a language.}
}
\value{
Returns a character vector containing a two-letter language code.
}
\description{
This is a lookup function that returns the two-letter language code for specified language.
}
\examples{
language_code("French")
}
