% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{language_graphs}
\alias{language_graphs}
\title{Generates a graph of languages to search}
\usage{
language_graphs(lang_data, no_return = 15, key_topics = c("biology",
  "conservation", "ecology"))
}
\arguments{
\item{lang_data}{The exported language data from choose_languages().}

\item{no_return}{The maximum number of languages to include in the graph.}

\item{key_topics}{The same key topics used in choose_languages().}
}
\description{
Creates a bubble graph showing the number of journals in each language from the languages used by journals in given topics.
}
