% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_extraction.R
\name{make_corpus}
\alias{make_corpus}
\title{Create a corpus from a data frame}
\usage{
make_corpus(df)
}
\arguments{
\item{df}{a data frame where at least one column is called 'text'}
}
\value{
a corpus object
}
\description{
Calls the corpus function from quanteda to create a corpus from the data frame with search hits.
}
