% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{make_dataset}
\alias{make_dataset}
\title{Generate co-occurrence dataset}
\usage{
make_dataset(df, keys, onlykeywords = FALSE)
}
\arguments{
\item{df}{The study keyword dataset.}

\item{keys}{The unique keywords from the sample.}

\item{onlykeywords}{If set to \code{TRUE}, the search is restricted to author-tagged keywords rather than the descriptors generated with \code{get_descriptors}.}
}
\value{
A presence-absence dataset of keywords for each study.
}
\description{
Takes the study keyword dataset and converts it to a matrix of 1s and 0s indicating presence (or absence) of a keyword in that study.
}
