% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_extraction.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Make a dictionary from keywords}
\usage{
make_dictionary(actual_terms = select_actual_terms(df, min_freq = 2),
  likely_terms = extract_terms(df, new_stopwords = NULL, min_freq = 2, title =
  TRUE, abstract = TRUE))
}
\arguments{
\item{actual_terms}{a character vector of search terms}

\item{likely_terms}{a character vector of search terms}
}
\value{
a quanteda dictionary object
}
\description{
Combines actual keywords and likely keywords into a dictionary object using the as.dictionary function from quanteda.
}
