% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graphs.R
\name{make_graph}
\alias{make_graph}
\title{Convert occurrence matrix to graph}
\usage{
make_graph(trimmed_data, printplot = TRUE)
}
\arguments{
\item{trimmed_data}{The presence-absence dataset that has been trimmed down to only include common keywords.}
}
\value{
An \code{igraph} graph object.
}
\description{
Converts the presence-absence dataset to a keyword co-occurrence network stored as an \code{igraph} object.
}
