% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{make_importance}
\alias{make_importance}
\title{Subset strength data from a graph}
\usage{
make_importance(graph, importance_method = "strength")
}
\arguments{
\item{graph}{an igraph graph}

\item{importance_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node strengths, ranks, and names
}
\description{
Selects only the node strength data from a graph.
}
\examples{
my_graph <- create_network(
  search_dfm = create_dfm(
    c(
      "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
      "Variation in home-range size of Black-backed Woodpeckers"
    ),
    features = c("occupancy", "variation", "woodpecker"),
    closure = "none"
  ),
  min_studies = 1,
  min_occurrences = 1
)

make_importance(my_graph, importance_method = "strength")
}
