% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{make_ngram_graph}
\alias{make_ngram_graph}
\title{Create keyword co-occurrence network with only ngrams}
\usage{
make_ngram_graph(graph, min_ngrams = 2, unigrams = FALSE)
}
\arguments{
\item{graph}{an igraph object}

\item{min_ngrams}{a number; the minimum number of words to consider an ngram}

\item{unigrams}{if TRUE, returns a subset of the network where each node is a unigram}
}
\value{
an igraph object
}
\description{
Reduces the full keyword co-occurrence network to only include nodes with 2+ words or only unigrams. This is useful for separating commonly used words from distinct phrases.
}
\examples{
make_ngram_graph(graph=litsearchr::BBWO_graph, min_ngrams=2, unigrams=FALSE)
}
