% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{make_strengths}
\alias{make_strengths}
\title{Subset strength data from a graph}
\usage{
make_strengths(graph)
}
\arguments{
\item{graph}{an igraph graph}
}
\value{
a data frame of node strengths, ranks, and names
}
\description{
Selects only the node strength data from a graph.
}
