% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plots.R
\name{make_wordle}
\alias{make_wordle}
\title{Create a wordcloud}
\usage{
make_wordle(graph, colorchoice = c("black"))
}
\arguments{
\item{graph}{The reduced graph of only important nodes.}

\item{colorchoice}{A vector listing colors to use for the wordcloud.}
}
\value{
Saves wordcloud to a .png file.
}
\description{
Makes a wordcloud of keywords identified as important and changes size of text to correspond to node strength.
}
