% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{merge_stems}
\alias{merge_stems}
\title{Merge terms that have identical stems}
\usage{
merge_stems(dfm)
}
\arguments{
\item{dfm}{An object of class simple_triplet_matrix.}
}
\value{
The input dfm with duplicated terms merged into one column.
}
\description{
This function merges terms in document feature matrices if they have identical lemmatized stems.
}
\examples{
my_df <-  data.frame(
  title = c(
    "EviAtlas: a tool for visualising evidence synthesis databases",

    "revtools: An R package to support article screening for evidence synthesis",

    "An automated approach to identifying search terms for
    systematic reviews using keyword co‐occurrence networks",

    "Reproducible, flexible and high‐throughput data extraction
    from primary literature: The metaDigitise r package"
  ),

  year = c("2019", "2019", "2019", "2019"),

  authors = c("Haddaway et al", "Westgate",
              "Grames et al", "Pick et al")
)


terms <- c("evidence synthesis", "package", "systematic")

create_dtm(as.character(my_df$title), ngram_check = TRUE, min_freq = 1)
}
