% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plots.R
\name{plot_3D_network}
\alias{plot_3D_network}
\title{Plot a co-occurrence network in 3D}
\usage{
plot_3D_network(graph, graphcolor = "#006E6D")
}
\arguments{
\item{graph}{an igraph object}

\item{graphcolor}{a valid color name or hex code for the graph nodes}
}
\value{
an RGL plot of the graph
}
\description{
Plots a keyword co-occurrence network as an interactive, 3D RGL object. Doesn't really serve any purpose other than seeing node density, but it looks cool. Node sizes are the square root of node strength; edge widths are fixed at .25. Depends on the rgl package.
}
