% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plots.R
\name{plot_full_network}
\alias{plot_full_network}
\title{Plot full keyword co-occurrence network}
\usage{
plot_full_network(graph, graphcolor = "black")
}
\arguments{
\item{graph}{an igraph object}

\item{graphcolor}{a color choice for the graph edges and node frames}
}
\value{
a plot of the full network structure
}
\description{
Passes plotting parameters to igraph to suppress node names and show the structure of the full network.
}
