% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{possible_langs}
\alias{possible_langs}
\title{Languages litsearchr can write searches in}
\format{A database with 53 rows of 4 variables:
\describe{
  \item{Short}{the short form language code}
  \item{Language}{the name of the language}
  \item{Encoding}{which character encoding to use for a language}
  \item{Used}{whether or not the language is used by a scientific journal}
}}
\usage{
possible_langs
}
\description{
A dataset of the languages that searches can be translated
into along with their short form, character encoding, and
whether a scientific journal indexed in ulrich uses them.
}
\keyword{datasets}
