% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{remove_punctuation}
\alias{remove_punctuation}
\title{Remove punctuation from text}
\usage{
remove_punctuation(text, preserve_punctuation = NULL)
}
\arguments{
\item{text}{A character vector from which to remove punctuation.}

\item{preserve_punctuation}{A string or vector of punctuation to retain}
}
\value{
Returns the input text with punctuation removed.
}
\description{
Removes common punctuation marks from a text.
}
\examples{
remove_punctuation("#s<<<//<y>!&^n$$t/>h\%e&s$is#!++r!//")
}
