% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{remove_redundancies}
\alias{remove_redundancies}
\title{Remove redundant terms}
\usage{
remove_redundancies(terms, closure = c("left", "right", "full", "none"))
}
\arguments{
\item{terms}{a character vector of terms}

\item{closure}{restrictions on how matches are detected; left requires matches to start with a term (e.g "burn" matches "burning"), right requires matches to end with a term (e.g. "burn" matches "postburn" but not "postburning"), full requires exact matches (e.g. "burn" only matches "burn"), and none allows terms to be embedded within matches.}
}
\value{
a character vector with redundant terms removed
}
\description{
Given a list of terms, removes redundant terms based on plurals, stemming, or partial matches
}
