% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_text_functions.R
\name{replace_ngrams}
\alias{replace_ngrams}
\title{Replace n-grams in text as single terms}
\usage{
replace_ngrams(x, ngrams)
}
\arguments{
\item{x}{A character vector in which to replace n-grams.}

\item{ngrams}{A character vector of n-grams.}
}
\value{
The input character vector with spaces in n-grams replaced by underscores.
}
\description{
This function replaces spaces in n-grams with underscores to turn them into single tokens.
}
\examples{
replace_ngrams("by means of natural selection", "natural selection")
}
