% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scopus_example}
\alias{scopus_example}
\title{An example of exported naive search hits}
\format{A dataframe with 10 rows of 22 variables:
\describe{
\item{Authors}{Article authors}
\item{Author.Ids}{Author identifications}
\item{Title}{Article title}
\item{Year}{Year of article publication}
\item{Source.title}{Journal or source title}
\item{Volume}{Volume of publication}
\item{Issue}{Issue of publication}
\item{Art..No.}{Article number in journal, if applicable}
\item{Page.start}{First page of article}
\item{Page.end}{Last page of article}
\item{Page.count}{Total number of pages in article}
\item{Cited.by}{Number of times article has been cited}
\item{DOI}{Article DOI}
\item{Link}{Link to article}
\item{Affiliations}{Author affiliations}
\item{Authors.with.affiliations}{Author affiliations associated with each author}
\item{Abstract}{Article abstract}
\item{Author.keywords}{Author-tagged keywords}
\item{Document.Type}{Type of article (e.g. journal article, book chapter)}
\item{Access.Type}{Whether or not the article is open access}
\item{Source}{The database source}
\item{EID}{The identifier within the source}
}}
\usage{
scopus_example
}
\description{
A subset of the results of a naive search in Scopus
about Black-backed Woodpeckers (Picoides arcticus).
}
\keyword{datasets}
