% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_ingenta}
\alias{scrape_ingenta}
\title{Scrapes results from Ingenta Connect}
\usage{
scrape_ingenta(search_terms, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\description{
Scrapes hits from Ingenta Connect. Query length is limited by server requests, which can be triggered either by excessively long queries or by vague queries that return too many results. Must login through institutional access.
}
