% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_jstor}
\alias{scrape_jstor}
\title{Scrapes results from JSTOR}
\usage{
scrape_jstor(search_terms, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\value{
a database of hits (if yes is selected from the menu prompt, the hits will also be saved to your working directory)
}
\description{
Scrapes hits from JSTOR. Limited to 256 characters and must access through institutional login.
}
