% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{scrape_ndltd}
\alias{scrape_ndltd}
\title{Scrapes results from the NDLTD Global ETD Search}
\usage{
scrape_ndltd(
  search_terms = NULL,
  URL = NULL,
  writefile = FALSE,
  verbose = TRUE,
  languages = "English",
  stemming = FALSE,
  exactphrase = TRUE,
  where = "description",
  directory = "./"
)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{URL}{the URL from conducting a search in the database}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}

\item{verbose}{if TRUE, prints which page of hits it has finished}

\item{languages}{which language to search in; available languages can be viewed with available_languages().}

\item{stemming}{if TRUE, keywords will be truncated and stem from root word forms (only if language is English)}

\item{exactphrase}{if TRUE, keyword phrases will be placed in quotes to ensure exact phrases are returned}

\item{where}{where in a thesis or dissertation to search (options are description or title)}

\item{directory}{the directory to save results to if writefile=TRUE}
}
\value{
a data frame containing the results of the search
}
\description{
Scrapes hits from the Networked Digital Library of Theses and Dissertations Global ETD Search.
}
\examples{
\dontrun{scrape_ndltd(search_terms=list(c("black backed woodpecker", "picoides arcticus")))}
}
