% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_oatd}
\alias{scrape_oatd}
\title{Scrapes results from Open Access Theses and Dissertations}
\usage{
scrape_oatd(search_terms, writefile = TRUE, verbose = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}

\item{verbose}{if TRUE, prints which page of hits it has finished}
}
\value{
a database of hits (if yes is selected from the menu prompt, the hits will also be saved to your working directory)
}
\description{
Scrapes hits from OATD. Query length is limited by server requests, which can be triggered either by excessively long queries.
}
