% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_pubmed}
\alias{scrape_pubmed}
\title{Scrapes results from PubMed}
\usage{
scrape_pubmed(search_terms, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\description{
Scrapes hits from PubMed. Query length is limited by server requests, which can be triggered either by excessively long queries.
}
