% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{scrape_results}
\alias{scrape_results}
\title{Scrapes results from open access thesis databases}
\usage{
scrape_results(URL, save_results = TRUE, destination = "./",
  verbose = TRUE)
}
\arguments{
\item{URL}{the URL from running a search}

\item{save_results}{if TRUE, writes results to a .csv file}

\item{destination}{the folder where the full search should be saved}
}
\description{
Scrapes web pages to download results into a datasheet that can be read with import_naive(). Currently only supports OATD, NDLTD, and OpenThesis.
}
