% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_scopus}
\alias{scrape_scopus}
\title{Scrapes results from Scopus}
\usage{
scrape_scopus(search_terms, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\description{
Scrapes hits from Scopus. Does not currently extract abstracts - only titles and publication information. Must login through institutional access.
}
