% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_WoS}
\alias{scrape_WoS}
\title{Scrapes results from Web of Science databases}
\usage{
scrape_WoS(sessionID, dbID = "UA", writefile = TRUE, qid = "1",
  verbose = TRUE)
}
\arguments{
\item{sessionID}{the session ID (found in the URL of an active sessioin as SID=)}

\item{dbID}{the database you are scraping results from; if you want all the databases your institution has access to, this is likely "UA" whereas specific databases like BIOSIS "BCI" can be called.}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}

\item{qid}{if you have run more than one query in a session, check your search history to change the query id}

\item{verbose}{if TRUE, prints percentage updates every now and then so you know it is still running}
}
\description{
Scrapes hits from databases hosted on Web of Science. The exported session must be left active while running the script and you must sign in through your institutional login.
}
