% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_testing.R
\name{search_performance}
\alias{search_performance}
\title{Get precision and recall of a search}
\usage{
search_performance(no_desired, no_hits, no_articles)
}
\arguments{
\item{no_desired}{the number of  articles that should be returned}

\item{no_hits}{the number of good hits that a search found (can be found with check_recall)}

\item{no_articles}{the total number of articles that a search found}
}
\value{
the sensitivity, precision, and number needed to process
}
\description{
Measures the performance of a search by precision (specificity) and recall (sensitivity).
}
