% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_extraction.R
\name{select_actual_terms}
\alias{select_actual_terms}
\title{Extract actual article keywords}
\usage{
select_actual_terms(df, min_freq = 2)
}
\arguments{
\item{df}{a data frame of search hits from import_scope}

\item{min_freq}{a number, the minimum occurrences to be included}
}
\value{
a character vector of keywords actually occurring in the search dataset
}
\description{
Extracts actual author-and-database tagged keywords.
}
