% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{select_ngrams}
\alias{select_ngrams}
\title{Subset n-grams from node names}
\usage{
select_ngrams(graph, n = 2, imp_method = "strength")
}
\arguments{
\item{graph}{an igraph object}

\item{n}{a minimum number of words in an n-gram}

\item{imp_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node names, strengths, and rank
}
\description{
Selects only nodes from a graph whose node names are at least n-grams, where n is the minimum number of words in the node name. The default n-gram is a 2+-gram, which captures potential keyword terms that are at least two words long. The reason for this is that unigrams (terms with only one word) are detected more frequently, but are also generally less relevant to finding keyword terms.
}
\examples{
my_graph <- create_network(
  search_dfm = create_dfm(
    c(
      "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
      "Variation in home-range size of Black-backed Woodpeckers",
      "Black-backed woodpecker occupancy in burned and beetle-killed forests"
    ),
    features = c("occupancy", "variation", "woodpecker", "burn"),
    closure = "none"
  ),
  min_studies = 1,
  min_occ = 1
)

select_ngrams(my_graph, imp_method = "strength")
}
