% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{select_unigrams}
\alias{select_unigrams}
\title{Subset unigrams from node names}
\usage{
select_unigrams(graph, importance_method = "strength")
}
\arguments{
\item{graph}{an igraph object}

\item{importance_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node names, strengths, rank
}
\description{
Selects only nodes from a graph whose node names are at single words.
}
\examples{
my_graph <- create_network(
  search_dfm = create_dfm(
    c(
      "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
      "Variation in home-range size of Black-backed Woodpeckers"
    ),
    features = c("occupancy", "variation", "woodpecker"),
    closure = "none"
  ),
  min_studies = 1,
  min_occurrences = 1
)

select_ngrams(my_graph, importance_method = "strength")
}
