% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{trim_dataset}
\alias{trim_dataset}
\title{Remove rare keywords}
\usage{
trim_dataset(dataset, studies = 2)
}
\arguments{
\item{dataset}{The presence-absence dataset of keywords.}

\item{studies}{The minimum number of studies a keyword must appear in to stay in the dataset.}
}
\value{
A reduced dataset that does not contain entries for rare keywords.
}
\description{
Trims the dataset down to remove keyword column vectors that are rarely used. The primary purpose for this function is to reduce the computation time for the co-occurrence matrix and associated graph.
}
