% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ulrich}
\alias{ulrich}
\title{Non-English scientific journal list}
\format{A database with 5100 rows of 6 variables:
\describe{
  \item{Title}{the journal title}
  \item{ISSN}{the journal ISSN}
  \item{SubjectCodes}{disciplines covered by the journal, as tagged by Ulrich's}
  \item{Country}{the country of publication for the journal}
  \item{Start.Year}{the first year of journal publication}
  \item{Language}{the primary language of articles published in the journal}
}}
\usage{
ulrich
}
\description{
A dataset of non-english scientific journals indexed in
Ulrich's Global Serials Directory.
}
\keyword{datasets}
