% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{usable_databases}
\alias{usable_databases}
\title{Print databases/platform exports that litsearchr can import or search in}
\usage{
usable_databases()
}
\description{
Prints a data frame of platforms, databases, and download methods that litsearchr recognizes
}
\examples{
usable_databases()
}
