% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{write_search}
\alias{write_search}
\title{Write Boolean searches}
\usage{
write_search(groupdata, languages = choose_languages(key_topics = c("biology",
  "conservation", "ecology"))[1:10, 1], exactphrase = FALSE)
}
\arguments{
\item{groupdata}{A list of character vectors, each of which is a concept group.}

\item{languages}{Which languages to write searches in. The default relies on searching a database of journals by discipline. All supported languages can be seen with available_languages().}

\item{exactphrase}{If set to \code{TRUE}, stemmed search terms with multiple words will be enclosed in quotes.}
}
\description{
Takes search terms grouped by concept group and writes Boolean searches in which terms within concept groups are separated by "OR" and concept groups are separated by "AND". Searches can be written in up to 53 languages, though the function defaults to only searching the top ten most used languages in a discipline using the choose_languages() function.
}
