% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{write_stemmed_search}
\alias{write_stemmed_search}
\title{Write search with truncated word stems}
\usage{
write_stemmed_search(groupdata, languages = "English", exactphrase = FALSE,
  directory = "./")
}
\arguments{
\item{groupdata}{a list of character vectors, each of which is a concept group}

\item{languages}{a character of the language in which to write stemmed searches; currently, only English is supported}

\item{exactphrase}{if set to \code{TRUE}, stemmed search terms with multiple words will be enclosed in quotes}

\item{directory}{the path to the directory where you want to save searches (defaults to current working directory)}
}
\description{
Truncates words to word stems and appends them with an asterisk as a wildcard character. Currently only supported for English.
}
