% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{add_stopwords}
\alias{add_stopwords}
\title{Add new stopwords to ignore}
\usage{
add_stopwords(new_stopwords)
}
\arguments{
\item{new_stopwords}{a character vector of new words to add}
}
\value{
an updated vector of custom stopwords to remove from text
}
\description{
Allows the user to add additional stopwords to the built-in English stopwords list.
}
\examples{
add_stopwords(c("19th century", "abiotic factors", "elsevier direct"))
}
