% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_grouped_keywords}
\alias{BBWO_grouped_keywords}
\title{Example BBWO document-feature matrix}
\format{A document-feature matrix with 775 documents and 2 features

Grouped keywords for the BBWO vignette

A dataset containing the keyword groupings used in the
vignette about Black-backed Woodpecker occupancy of
post-fire forest systems.}
\usage{
BBWO_grouped_keywords
}
\description{
A document-feature matrix containing two example terms
from the vignette dictionary and imported naive search.
}
\keyword{datasets}
