% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BBWO_similar_grouped}
\alias{BBWO_similar_grouped}
\title{Similar terms from example}
\format{A database with 922 rows of 2 variables
\describe{
\item{group} the group we assigned
\item{term} the similar search term}}
\usage{
BBWO_similar_grouped
}
\description{
A dataset of terms marked as similar when we were
building an example search strategy.
}
\keyword{datasets}
