% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_testing.R
\name{check_recall}
\alias{check_recall}
\title{Check the recall of a search strategy}
\usage{
check_recall(true_hits, retrieved_articles, min_sim = 0.6,
  new_stopwords = NULL)
}
\arguments{
\item{true_hits}{a character vector of titles for articles that should be returned}

\item{retrieved_articles}{a character vector of titles for articles returned by a search}

\item{min_sim}{the minimum similarity between two titles to be considered for manual review}

\item{new_stopwords}{any common words in the titles that should be ignored when computing similarity to avoid false matches}
}
\value{
a table of the best match for each true title from the search results along with a title similarity score
}
\description{
Checks a list of known articles against the results of a search to see how many the search retrieves.
}
