% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{choose_languages}
\alias{choose_languages}
\title{Select non-English languages to search in}
\usage{
choose_languages(key_topics = c("biology", "conservation", "ecology"))
}
\arguments{
\item{key_topics}{A vector of topics related to the topic of the search.}
}
\value{
A data frame of languages used by journals tagged with the key topics and a count of how many journals use that language.
}
\description{
This function chooses the best non-English languages to conduct searches in based on the topic of the review. The topics query a database of non-English language journals compiled from Ulrich; currently only STEM fields are supported.
}
