% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{clean_keywords}
\alias{clean_keywords}
\title{Remove duplicate studies and punctuation}
\usage{
clean_keywords(keywords)
}
\arguments{
\item{keywords}{a character vector containing keywords to clean}
}
\value{
a data frame with keyword punctuation standardized
}
\description{
Replaces all miscellaneous punctuation marks used to separate keywords and replaces them with a semicolon so that keywords properly separate in later steps.
}
\examples{
terms <- c("Picoides arcticus [Breeding season /  / ] [Nests /  / ]",
"[Forest and and woodland / Coniferous forest / ]",
"[Fire / Extreme wildfire / ] and [California / Sierra Nevada Mountains / ].")

clean_keywords(terms)

extract_terms(keywords=terms, method="tagged", min_freq = 1, ngrams = FALSE)
}
