% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{clean_keywords}
\alias{clean_keywords}
\title{Clean and standardize keyword punctuation}
\usage{
clean_keywords(keywords)
}
\arguments{
\item{keywords}{a character vector of keywords}
}
\value{
a character vector of keywords with standardized punctuation
}
\description{
Replaces all miscellaneous punctuation marks used to separate keywords and replaces them with a semicolon so that keywords properly separate in later steps.
}
\examples{
clean_keywords(BBWO_data$keywords)
}
