% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_extraction.R
\name{create_dfm}
\alias{create_dfm}
\title{Create document-term matrix}
\usage{
create_dfm(corpus = make_corpus(df), my_dic = make_dictionary(),
  custom_stopwords = add_stopwords(NULL))
}
\arguments{
\item{corpus}{a corpus object}

\item{my_dic}{a dictionary object}

\item{custom_stopwords}{a character vector of words to ignore}
}
\value{
a quanteda dfm object
}
\description{
Calls the dfm function from quanteda to create a document-term matrix of terms included in a custom dictionary of potential terms.
}
