% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{create_dfm}
\alias{create_dfm}
\title{Create a document-feature matrix}
\usage{
create_dfm(elements, features, closure = "full")
}
\arguments{
\item{elements}{a character vector of document information (e.g. document titles or abstracts)}

\item{type}{whether the dfm should be created based on document tokens or a restricted list of keywords}

\item{language}{if type="tokens", the language to use for removing stopwords}

\item{keywords}{if type="keywords", a character vector of keywords to use as document features}
}
\value{
a matrix with documents as rows and terms as columns

a matrix with documents as rows and terms as columns
}
\description{
Given a character vector of document information, creates a document-feature matrix.
}
