% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{create_dfm}
\alias{create_dfm}
\title{Create a document-feature matrix}
\usage{
create_dfm(elements, features)
}
\arguments{
\item{elements}{a character vector of document information (e.g. document titles or abstracts)}

\item{features}{a character vector of terms to use as document features (e.g. keywords)}
}
\value{
a matrix with documents as rows and terms as columns
}
\description{
Given a character vector of document information, creates a document-feature matrix.
}
\examples{
dfm <- create_dfm(
  elements = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  features = c("occupancy", "variation", "black-backed woodpecker", "burn")
)

as.matrix(dfm)
}
