% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{deduplicate}
\alias{deduplicate}
\title{Remove duplicate articles}
\usage{
deduplicate(df, use_abstracts = TRUE, use_titles = TRUE,
  method = c("quick", "levenshtein", "tokens"), doc_sim = 0.85,
  title_sim = 0.95)
}
\arguments{
\item{df}{a data frame created with import_scope()}

\item{use_abstracts}{if TRUE, computes similarity scores for all abstracts}

\item{use_titles}{if TRUE, computes similarity based on titles}

\item{method}{if "tokens", tokenizes titles to compute similarity; if "quick", removes punctuation and capitalization then removes exact duplicates; if "levenshtein", uses levenshtein distances to compute similarity}

\item{doc_sim}{the minimum similarity between two abstracts to be marked as duplicated}

\item{title_sim}{the minimum similarity between two titles to be marked as duplicated}
}
\value{
a data frame with duplicates removed
}
\description{
Uses similarity of abstracts and titles to detect duplicates and remove them from the dataset.
}
\examples{
deduplicate(BBWO_import, use_abstracts=FALSE, use_titles=TRUE, method="tokens", title_sim=.95)
}
