% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract potential keywords from abstracts and titles}
\usage{
extract_terms(df, type = c("RAKE", "tagged"), new_stopwords = NULL,
  min_freq = 2, title = TRUE, abstract = TRUE, ngrams = TRUE,
  n = 2)
}
\arguments{
\item{df}{a dataframe created with import_scope}

\item{type}{if generating keywords from title and/or abstract, use "RAKE"; if extracting author or database tagged keywords, use "tagged"}

\item{new_stopwords}{a character vector of stopwords to ignore}

\item{min_freq}{a number, the minimum occurrences of a potential term}

\item{title}{include titles if TRUE}

\item{abstract}{include abstracts if TRUE}

\item{ngrams}{if TRUE, only extracts phrases with word count greater than a specified n}

\item{n}{the minimum word count for ngrams}
}
\value{
a character vector of potential keyword terms
}
\description{
Uses the RAKE (Rapid Automatic Keyword Extractor) from rapidraker to extract potential keyword terms from titles and abstracts.
}
\examples{
extract_terms(df=BBWO_data, type="RAKE")
}
