% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract potential keywords from abstracts and titles}
\usage{
extract_terms(text = NULL, keywords = NULL, method = c("fakerake",
  "RAKE", "tagged"), min_freq = 2, ngrams = TRUE, n = 2,
  language = "English")
}
\arguments{
\item{text}{A character object of text from which to extract terms}

\item{keywords}{A character vector of keywords tagged by authors and/or databases if using method="tagged"}

\item{method}{The method of extracting keywords; options are fakerake (a quick implementation similar to Rapid Automatic Keyword Extraction), RAKE, or tagged for author-tagged keywords}

\item{min_freq}{a number, the minimum occurrences of a potential term}

\item{ngrams}{if TRUE, only extracts phrases with word count greater than a specified n}

\item{n}{the minimum word count for ngrams}

\item{language}{the language of input data to use for stopwords}
}
\value{
a character vector of potential keyword terms
}
\description{
Extracts potential keyword terms from text (e.g. titles and abstracts)
}
\examples{
extract_terms(df=BBWO_data, type="RAKE")
}
