% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_extraction.R
\name{extract_terms}
\alias{extract_terms}
\title{Extract potential keywords from abstracts and titles}
\usage{
extract_terms(df, new_stopwords = NULL, min_freq = 2, title = TRUE,
  abstract = TRUE)
}
\arguments{
\item{df}{a dataframe created with import_scope}

\item{new_stopwords}{a character vector of stopwords to ignore}

\item{min_freq}{a number, the minimum occurrences of a potential term}

\item{title}{include titles if TRUE}

\item{abstract}{include abstracts if TRUE}
}
\value{
a character vector of potential keyword terms
}
\description{
Uses the RAKE (Rapid Automatic Keyword Extractor) from rapidraker to extract potential keyword terms from titles and abstracts.
}
