% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{fakerake}
\alias{fakerake}
\title{Quick keyword extraction}
\usage{
fakerake(text, stopwords)
}
\arguments{
\item{text}{A string object to extract terms from}

\item{stopwords}{A character vector of stopwords to remove}
}
\value{
A character vector of potential keywords
}
\description{
Extracts potential keywords from text separated by stopwords
}
