% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{find_cutoff}
\alias{find_cutoff}
\title{Find node cutoff strength}
\usage{
find_cutoff(graph, method = c("spline", "cumulative"), cum_pct = 0.8,
  degrees = 2, knot_num = 1, diagnostics = TRUE)
}
\arguments{
\item{graph}{The complete graph.}

\item{method}{The spline fit finds tipping points in the ranked order of node strengths to use as cutoffs. The cumulative fit option finds the node strength cutoff point at which a certain percent of the total strength of the graph is captured (e.g. the fewest nodes that contain 80\% of the total strength).}

\item{cum_pct}{if using method cumulative, the total percent of node strength to capture}

\item{degrees}{if using method spline, the degrees of the polynomial curve that approximates the ranked unique node strengths}

\item{knot_num}{if using method spline, the number of knots to allow}

\item{diagnostics}{if set to TRUE, saves plots of either the fit splines and residuals or the curve of cumulative node strength and cutoff point}
}
\value{
a vector of suggested node cutoff strengths
}
\description{
Find the minimum node strength to use as a cutoff point for important nodes.
}
