% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{fit_splines}
\alias{fit_splines}
\title{Fit spline model to node strengths}
\usage{
fit_splines(strength_data, degrees = 2, knot_num = 1, knots)
}
\arguments{
\item{strength_data}{a dataset of ranked unique node strengths}

\item{degrees}{the same degrees used to find knot placement in \code{find_knots}}

\item{knot_num}{the same number of knots used to find knot placement in \code{find_knots}}

\item{knots}{The vector of optimal knots returned from \code{find_knots}}
}
\value{
a fitted spline model
}
\description{
Fits a basis spline to the curve of ranked unique node strengths.
}
