% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{get_keywords}
\alias{get_keywords}
\title{Extract potential keywords}
\usage{
get_keywords(reduced_graph, savekeywords = FALSE, makewordle = FALSE,
  directory = "./")
}
\arguments{
\item{reduced_graph}{a reduced graph with only important nodes created with reduce_grah()}

\item{savekeywords}{if TRUE, saves the keywords to a plain text file}

\item{makewordle}{if TRUE, creates a wordcloud image of the important keywords sized relative to node strength}

\item{directory}{the directory to save results to if savekeywords=TRUE}
}
\value{
a list of potential keywords to consider
}
\description{
Extracts keywords identified as important.
}
\examples{
get_keywords(reduce_graph(litsearchr::BBWO_graph, cutoff_strength=15))
}
