% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_selection.R
\name{get_keywords}
\alias{get_keywords}
\title{Extract potential keywords}
\usage{
get_keywords(reduced_graph, savekeywords = TRUE, makewordle = TRUE)
}
\arguments{
\item{reduced_graph}{The reduced graph with only important nodes.}

\item{savekeywords}{If TRUE, saves the keywords to a plain text file.}

\item{makewordle}{If TRUE, creates a wordcloud image of the important keywords sized by node strength.}
}
\value{
The list of potential keywords to consider.
}
\description{
Extracts keywords identified as important and writes them to a plain text file.
}
