% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{get_keywords}
\alias{get_keywords}
\title{Extract potential keywords}
\usage{
get_keywords(reduced_graph)
}
\arguments{
\item{reduced_graph}{a reduced graph with only important nodes created with reduce_grah()}
}
\value{
a character vector of potential keywords to consider
}
\description{
Extracts keywords identified as important.
}
\examples{
get_keywords(reduce_graph(litsearchr::BBWO_graph, cutoff_strength=15))
}
