% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{get_keywords}
\alias{get_keywords}
\title{Extract potential keywords}
\usage{
get_keywords(reduced_graph)
}
\arguments{
\item{reduced_graph}{a reduced graph with only important nodes created with reduce_grah()}
}
\value{
a character vector of potential keywords to consider
}
\description{
Extracts keywords identified as important.
}
\examples{
dfm <- create_dfm(
  elements = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  features = c("occupancy", "variation", "black-backed woodpecker", "burn")
)

my_network <- create_network(search_dfm = as.matrix(dfm),
                             min_studies = 1,
                             min_occ = 1)

get_keywords(my_network)
}
