% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{import_results}
\alias{import_results}
\title{Import results of a search}
\usage{
import_results(directory = NULL, file = NULL, verbose = TRUE)
}
\arguments{
\item{directory}{a path to a directory containing search results to import}

\item{verbose}{if TRUE, prints status updates}

\item{filename}{a file of search results to import}
}
\value{
a data frame of assembled search results
}
\description{
Given a file or directory, calls the import_results function from the synthesisr package to import and assemble search results
}
