% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{import_results}
\alias{import_results}
\title{Import results of a search}
\usage{
import_results(directory = NULL, filename = NULL,
  save_dataset = FALSE, verbose = TRUE, save_directory = "./")
}
\arguments{
\item{directory}{a path to a directory containing search results to import}

\item{filename}{a path to a filename containing search results to import}

\item{save_dataset}{if TRUE, saves the full search results to a .csv}

\item{verbose}{if TRUE, prints status updates}

\item{save_directory}{the path to a directory where search results will be saved if save_dataset is set to TRUE}
}
\value{
a data frame of assembled search results
}
\description{
Given a file or directory, calls the import_results function from the synthesisr package to import and assemble search results
}
