% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{import_results}
\alias{import_results}
\title{Import results of a search}
\usage{
import_results(directory, remove_duplicates = FALSE,
  duplicate_methods = c("tokens", "quick", "levenshtein"),
  clean_dataset = TRUE, save_full_dataset = FALSE, verbose = TRUE,
  save_directory = "./")
}
\arguments{
\item{directory}{the full path to the directory in which the searches are saved}

\item{remove_duplicates}{if TRUE, removes duplicates based on document similarity}

\item{duplicate_methods}{the method to pass to deduplicate() if remove_duplicates is TRUE}

\item{clean_dataset}{if TRUE, removes excess punctuation and standardizes keywords}

\item{save_full_dataset}{if TRUE, saves a .csv of the full dataset in the working directory}

\item{verbose}{if TRUE, prints which file is currently being imported}

\item{save_directory}{the directory to save results to if save_full_dataset=TRUE}
}
\value{
a data frame of all the search results combined
}
\description{
Imports the results of a search, combines them into a single dataset, and (optionally) removes duplicate hits based on document similarity. Duplicates can be removed subsequently with custom similarity cutoffs using deduplicate() on the full dataset.
}
\examples{
write.csv(litsearchr::scopus_example, paste(tempdir(), "/scopus.csv", sep=""))
naive_results <- import_results(paste(tempdir(), "/", sep=""))

}
