% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{language_graphs}
\alias{language_graphs}
\title{Generates a graph of languages to search}
\usage{
language_graphs(lang_data = get_language_data(key_topics = NULL),
  no_return = 15, key_topics = NULL)
}
\arguments{
\item{lang_data}{a table of language data returned from choose_languages()}

\item{no_return}{the maximum number of languages to include in the graph}

\item{key_topics}{a character vector of the same key topics used in choose_languages()}
}
\value{
a bubble plot of non-English languages used by journals in a discipline sized by count
}
\description{
Creates a bubble graph showing the number of journals in each language from the languages used by journals in given topics.
}
\examples{
language_graphs(
  lang_data=get_language_data(key_topics=c("biology", "conservation", "ecology")),
  no_return=15,
  key_topics=c("biology", "conservation", "ecology"))
}
