% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Make a dictionary from keywords}
\usage{
make_dictionary(terms = NULL)
}
\arguments{
\item{terms}{a list object including at least one character vector of search terms}
}
\value{
a quanteda dictionary object
}
\description{
Combines actual keywords and likely keywords into a dictionary object using the as.dictionary function from quanteda.
}
\examples{
make_dictionary(terms=extract_terms(BBWO_data, type="RAKE", min_freq=3))
}
