% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{make_importance}
\alias{make_importance}
\title{Subset strength data from a graph}
\usage{
make_importance(graph, imp_method = "strength")
}
\arguments{
\item{graph}{an igraph graph}

\item{imp_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node strengths, ranks, and names
}
\description{
Selects only the node strength data from a graph.
}
\examples{
dfm <- create_dfm(
  elements = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  features = c("occupancy", "variation", "black-backed woodpecker", "burn")
)

my_network <- create_network(search_dfm = as.matrix(dfm),
                             min_studies = 1,
                             min_occ = 1)

make_importance(my_network, imp_method = "strength")
}
