% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{make_importance}
\alias{make_importance}
\title{Subset strength data from a graph}
\usage{
make_importance(graph, importance_method)
}
\arguments{
\item{graph}{an igraph graph}

\item{importance_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node strengths, ranks, and names
}
\description{
Selects only the node strength data from a graph.
}
