% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_selection.R
\name{make_strengths}
\alias{make_strengths}
\title{Find node strength and rank}
\usage{
make_strengths(graph)
}
\arguments{
\item{graph}{An \code{igraph} graph object.}
}
\value{
Dataset of unique node strengths and in ranked order.
}
\description{
Creates a dataset of graph node strengths and ranks.
}
