% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plots.R
\name{make_wordle}
\alias{make_wordle}
\title{Create a wordcloud}
\usage{
make_wordle(graph, colorchoice = c("#006E6D"))
}
\arguments{
\item{graph}{a reduced graph of only important nodes}

\item{colorchoice}{a vector listing colors to use for the wordcloud}
}
\value{
plots a wordcloud
}
\description{
Makes a wordcloud of keywords identified as important and changes size of text to correspond to node strength.
}
\examples{
make_wordle(reduce_graph(litsearchr::BBWO_graph, 15, importance_method="strength"))
}
