% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_plots.R
\name{plot_full_network}
\alias{plot_full_network}
\title{Plot full keyword co-occurrence network}
\usage{
plot_full_network(graph, graphcolor = "#006E6D", color_gradient = TRUE)
}
\arguments{
\item{graph}{an igraph object}

\item{graphcolor}{a color choice for the graph edges and node frames}

\item{color_gradient}{if TRUE, the intensity of the color for each node scales with node strength}
}
\value{
a plot of the full network structure
}
\description{
Passes plotting parameters to igraph to suppress node names and show the structure of the full network.
}
\examples{
plot_full_network(reduce_graph(litsearchr::BBWO_graph, 15, importance_method="strength"))
}
