% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{reduce_graph}
\alias{reduce_graph}
\title{Create reduced graph of important nodes
Takes the full graph and reduces it to only include nodes (and associated edges) greater than the cutoff strength for important nodes.}
\usage{
reduce_graph(graph, cutoff_strength, importance_method = "strength")
}
\arguments{
\item{graph}{the full graph object}

\item{cutoff_strength}{the minimum node importance to be included in the reduced graph}

\item{importance_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
an igraph graph with only important nodes
}
\description{
Create reduced graph of important nodes
Takes the full graph and reduces it to only include nodes (and associated edges) greater than the cutoff strength for important nodes.
}
\examples{
my_graph <- create_network(
  search_dfm = create_dfm(
    c(
      "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
      "Variation in home-range size of Black-backed Woodpeckers",
      "Black-backed woodpecker occupancy in burned and beetle-killed forests"
    ),
    features = c("occupancy", "variation", "woodpecker"),
    closure = "none"
  ),
  min_studies = 1,
  min_occurrences = 1
)
create_network(
  search_dfm = create_dfm(
    c(
      "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
      "Variation in home-range size of Black-backed Woodpeckers",
      "Black-backed woodpecker occupancy in burned and beetle-killed forests"
    ),
    features = c("occupancy", "variation", "woodpecker"),
    closure = "none"
  ),
  min_studies = 1,
  min_occurrences = 1
)

reduce_graph(my_graph, cutoff_strength = 1)
}
