% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graphs.R
\name{reduce_graph}
\alias{reduce_graph}
\title{Create reduced graph of important nodes
Takes the full graph and reduces it to only include nodes (and associated edges) greater than the cutoff strength for important nodes.}
\usage{
reduce_graph(graph, cutoff_strength, printplot = TRUE)
}
\arguments{
\item{graph}{The full graph object.}

\item{cutoff_strength}{The minimum node strength to be included in the reduced graph.}
}
\value{
An \code{igraph} graph with only important nodes.
}
\description{
Create reduced graph of important nodes
Takes the full graph and reduces it to only include nodes (and associated edges) greater than the cutoff strength for important nodes.
}
