% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_clean_data.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate entries from a data frame}
\usage{
remove_duplicates(df, field, method = c("string_osa", "fuzzdist", "exact"))
}
\arguments{
\item{df}{the data frame to deduplicate}

\item{field}{the name or index of the column to check for duplicate values}

\item{method}{the manner of duplicate detection; exact removes exact text duplicates, stringdist removes duplicates below a similarity threshold, and fuzzy uses fuzzdist matching}
}
\value{
a deduplicated data frame
}
\description{
Calls the deduplicate function from synthesisr to flag and remove duplicate entries from a data frame
}
\examples{
my_df <- data.frame(title=c("Picoides", "picoides", "Seiurus"), id=c("01", "02", "03"))
remove_duplicates(my_df, "title", "exact")
}
