% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_google_scholar}
\alias{scrape_google_scholar}
\title{Scrapes results from Google Scholar}
\usage{
scrape_google_scholar(search_terms, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\description{
Scrapes hits from Google Scholar. Limited to 256 characters and 60 queries.
}
