% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{scrape_hits}
\alias{scrape_hits}
\title{Scrapes hits from specified databases}
\usage{
scrape_hits(
  search_terms = NULL,
  URL = NULL,
  database = c("oatd", "ndltd", "openthesis"),
  verbose = TRUE,
  writefile = FALSE,
  directory = "./"
)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms.}

\item{URL}{the URL from searching in OATD, NDLTD, or OpenThesis}

\item{database}{a character with the database to scrape (either "ndltd", "openthesis", or "oatd").}

\item{verbose}{if TRUE, prints which page of hits it has finished}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}

\item{directory}{the directory to save results to if writefile=TRUE}
}
\value{
a database of hits (if yes is selected from the menu prompt, the hits will also be saved to your working directory)
}
\description{
Provides a wrapper function to scrape hits from databases for which litsearchr can recognize formats.
}
