% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_results.R
\name{scrape_hits}
\alias{scrape_hits}
\title{Scrapes hits from specified databases}
\usage{
scrape_hits(search_terms, database = c("scopus", "googlescholar",
  "jstor", "cabdirect", "WoS", "ingenta", "pubmed", "worldcat", "oatd",
  "ndltd", "openthesis"), verbose = TRUE, writefile = TRUE)
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms. Note that for some databases (e.g. JSTOR) the query length is limited.}

\item{database}{a character with the database to scrape. Options: scopus, googlescholar, jstor, cabdirect, ingenta, pubmed, worldcat, oatd, ndltd, and openthesis.}

\item{verbose}{if TRUE, prints which page of hits it has finished}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}
}
\value{
a database of hits (if yes is selected from the menu prompt, the hits will also be saved to your working directory)
}
\description{
Provides a wrapper function to scrape hits from databases that litsearchr can scrape and that the user has institutional access to.
}
