% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_scrape_test_searches.R
\name{scrape_oatd}
\alias{scrape_oatd}
\title{Scrapes results from Open Access Theses and Dissertations}
\usage{
scrape_oatd(search_terms = NULL, URL = NULL, writefile = FALSE,
  verbose = TRUE, languages = "English", stemming = FALSE,
  exactphrase = TRUE, directory = "./")
}
\arguments{
\item{search_terms}{a list of character strings with grouped search terms}

\item{URL}{the URL from conducting a search in the database}

\item{writefile}{if TRUE, writes results to a .csv file in the working directory}

\item{verbose}{if TRUE, prints which page of hits it has finished}

\item{languages}{which language to search in; available languages can be viewed with available_languages().}

\item{stemming}{if TRUE, keywords will be truncated and stem from root word forms (only if language is English)}

\item{exactphrase}{if TRUE, keyword phrases will be placed in quotes to ensure exact phrases are returned#' @return a database of hits (if yes is selected from the menu prompt, the hits will also be saved to your working directory)}

\item{directory}{the directory to save results to if writefile=TRUE}
}
\value{
a data frame containing the results of the search
}
\description{
Scrapes hits from OATD. Query length is limited by server requests, which can be triggered either by excessively long queries.
}
\examples{
scrape_oatd(search_terms=list(c("black-backed woodpecker", "picoides arcticus")))
}
