% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{select_ngrams}
\alias{select_ngrams}
\title{Subset n-grams from node names}
\usage{
select_ngrams(graph, n = 2)
}
\arguments{
\item{graph}{an igraph object}

\item{n}{a minimum number of words in an n-gram}
}
\value{
a data frame of node names, strengths, rank
}
\description{
Selects only nodes from a graph whose node names are at least n-grams, where n is the minimum number of words in the node name. The default n-gram is a 2+-gram, which captures potential keyword terms that are at least two words long. The reason for this is that unigrams (terms with only one word) are detected more frequently, but are also generally less relevant to finding keyword terms.
}
