% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCN_node_selection.R
\name{select_unigrams}
\alias{select_unigrams}
\title{Subset unigrams from node names}
\usage{
select_unigrams(graph, importance_method = "strength")
}
\arguments{
\item{graph}{an igraph object}

\item{importance_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node names, strengths, rank
}
\description{
Selects only nodes from a graph whose node names are at single words.
}
