% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_selection.R
\name{select_unigrams}
\alias{select_unigrams}
\title{Subset unigrams from node names}
\usage{
select_unigrams(graph, imp_method = "strength")
}
\arguments{
\item{graph}{an igraph object}

\item{imp_method}{a character specifying the importance measurement to be used; takes arguments of "strength", "eigencentrality", "alpha", "betweenness", "hub" or "power"}
}
\value{
a data frame of node names, strengths, and rank
}
\description{
Selects only nodes from a graph whose node names are single words.
}
\examples{
dfm <- create_dfm(
  elements = c(
    "Cross-scale occupancy dynamics of a postfire specialist
    in response to variation across a fire regime",
    "Variation in home-range size of Black-backed Woodpeckers",
    "Black-backed woodpecker occupancy in burned and beetle-killed forests"
  ),
  features = c("occupancy", "variation", "black-backed woodpecker", "burn")
)

my_network <- create_network(
  search_dfm = as.matrix(dfm),
  min_studies = 1,
  min_occ = 1
)

select_ngrams(my_network, imp_method = "strength")
select_unigrams(my_network, imp_method = "strength")
}
