% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{should_stem}
\alias{should_stem}
\title{Check whether a word is long enough to stem}
\usage{
should_stem(word)
}
\arguments{
\item{word}{the word to check}
}
\description{
Checks if the stemmed form of a word is longer than 3 characters. Not intended as a standalone function and is called from write_stemmed_search().
}
