% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{translate_search}
\alias{translate_search}
\title{Translate search terms}
\usage{
translate_search(search_terms, target_language, source_language = "en",
  API_key = "AIzaSyBNlOXp3YIbUw2BOUqMAK4hNf5-t5EU6wY")
}
\description{
Takes groups of search terms and translates them into target language using the Google Translate API. This function is intended for use inside write_search(), not as a standalone function.
}
