% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{write_stemmed_search}
\alias{write_stemmed_search}
\title{Write search with truncated word stems}
\usage{
write_stemmed_search(groupdata, languages = "English", exactphrase = FALSE)
}
\arguments{
\item{groupdata}{A list of character vectors, each of which is a concept group.}

\item{languages}{Currently, only English stemmed searches are supported.}

\item{exactphrase}{If set to \code{TRUE}, stemmed search terms with multiple words will be enclosed in quotes.}
}
\description{
Truncates words to word stems and appends them with an asterisk as a wildcard character. Currently only supported for English.
}
