% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_searches.R
\name{write_title_search}
\alias{write_title_search}
\title{Write a search to check title recall}
\usage{
write_title_search(titles)
}
\arguments{
\item{titles}{a character vector of titles}
}
\value{
a text string
}
\description{
Given a set of titles, writes a Boolean search that can be used in database title fields to check whether a search strategy retrieves titles.
}
